requirements.txt
numpy==1.24.4
pillow==11.3.0
matplotlib==3.10.3
tensorflow==2.10.1
scikit-learn==1.5.1
streamlit==1.22.0
seaborn==0.13.2
opencv-python==4.10.0.84
